moving methods for get_traces, get_image,pixel_masks, num_of_frames/rois, get_roi_locations to base class
