+++ roiextractors/extractors/suite2p/suite2psegmentationextractor.py
@@ -35,6 +35,7 @@ class Suite2pSegmentationExtractor(SegmentationExtractor):
         self.iscell = self._load_npy('iscell.npy', mmap_mode='r')
         self.ops = self._load_npy('ops.npy').item()
         self._channel_names = [f'OpticalChannel{i}' for i in range(self.ops['nchannels'])]
+        self._roi_response = self.F
         self._roi_response_dict = {'Fluorescence': self.F,
                                'Neuropil': self.Fneu,
                                'Deconvolved': self.spks}
@@ -63,41 +64,6 @@ class Suite2pSegmentationExtractor(SegmentationExtractor):
     # defining the abstract class enforced methods:
     def get_roi_ids(self):
         return list(range(self.no_rois))
-
-    def get_num_rois(self):
-        return len(self.stat)
-
-    def get_roi_locations(self, roi_ids=None):
-        if roi_ids is None:
-            return self.roi_locations
-        else:
-            roi_idx = [np.where(np.array(i) == self.roi_ids)[0] for i in roi_ids]
-            ele = [i for i, j in enumerate(roi_idx) if j.size == 0]
-            roi_idx_ = [j[0] for i, j in enumerate(roi_idx) if i not in ele]
-            return self.roi_locations[:, roi_idx_]
-
-    def get_num_frames(self):
-        return self.ops['nframes']
-
-    def get_traces(self, roi_ids=None, start_frame=None, end_frame=None, name='Fluorescence'):
-        if start_frame is None:
-            start_frame = 0
-        if end_frame is None:
-            end_frame = self.get_num_frames() + 1
-        if roi_ids is None:
-            roi_idx_ = range(self.get_num_rois())
-        else:
-            roi_idx = [np.where(np.array(i) == self.roi_ids)[0] for i in roi_ids]
-            ele = [i for i, j in enumerate(roi_idx) if j.size == 0]
-            roi_idx_ = [j[0] for i, j in enumerate(roi_idx) if i not in ele]
-        if name == 'Fluorescence':
-            return self.F[[roi_idx_], start_frame:end_frame]
-        if name == 'Neuropil':
-            return self.Fneu[[roi_idx_], start_frame:end_frame]
-        if name == 'Deconvolved':
-            return self.spks[[roi_idx_], start_frame:end_frame]
-        else:
-            return None
 
     def get_roi_image_masks(self, roi_ids=None):
         if roi_ids is None:
